% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{cnn_melanoma}
\alias{cnn_melanoma}
\title{CNN for melanoma data}
\usage{
cnn_melanoma(
  output_shape = NULL,
  mbl = FALSE,
  ll_activation = "linear",
  dropout_rate = 0.3,
  ll_bias = FALSE,
  input_shape = dim(im)[2:5],
  ...
)
}
\description{
CNN for melanoma data
}
