% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{cnn_mnist}
\alias{cnn_mnist}
\title{CNN for MNIST data}
\usage{
cnn_mnist(
  output_shape = NULL,
  mbl = FALSE,
  ll_activation = "linear",
  ll_bias = FALSE,
  input_shape = dim(im)[2:4],
  ...
)
}
\description{
CNN for MNIST data
}
