% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{cnn_utkface}
\alias{cnn_utkface}
\title{CNN for UTKFace data}
\usage{
cnn_utkface(
  output_shape = NULL,
  mbl = FALSE,
  ll_activation = "linear",
  dropout_rate = 0.3,
  ll_bias = FALSE,
  input_shape = dim(im)[2:4],
  ...
)
}
\description{
CNN for UTKFace data
}
