% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_randomsplits.R
\name{fit_randomsplits}
\alias{fit_randomsplits}
\title{Perform random split experiment}
\usage{
fit_randomsplits(
  mod = c("silscs", "sics", "cils", "ci", "si", "sils"),
  fml,
  tab_dat,
  im = NULL,
  ridx,
  splits = 6,
  nn = NULL,
  input_shape = NULL,
  bs = 1,
  lr = 5e-05,
  optimizer = optimizer_adam(learning_rate = lr),
  epochs = 10,
  loss = c("nll", "rps"),
  ws = FALSE,
  augment = FALSE,
  aug_params = list(rotation_range = 20, width_shift_range = 0.2, height_shift_range =
    0.2, zoom_range = 0.15, shear_range = 0.15, fill_mode = "nearest"),
  train_batchwise = FALSE,
  out_dir,
  fname
)
}
\arguments{
\item{fml}{model formula (e.g. \code{y ~ 1} for fitting a CI model).}

\item{tab_dat}{data frame of tabular data (including response variable). Response variable needs to be ordered
(also for binary responses).}

\item{ridx}{row indices used for each split (output from \code{get_ridx}).}

\item{nn}{function to build neural network used for modeling complex intercept or complex shift term.
Arguments \code{input_shape}, \code{output_shape} and \code{mbl} are required (see \code{\link{cnn_stroke}} for an example).}

\item{ws}{logical. Whether to initialize intercepts (simple and complex) and linear shift terms using the parameters estimated by a
\code{Polr} or \code{glm} model (depending on the response type).}

\item{augment}{logical. Whether to perform 2D image augmentation during training.}

\item{aug_params}{parameters used for image augmentation.}

\item{train_batchwise}{logical. If \code{TRUE} model is trained batchwise (used to save memory).}

\item{out_dir}{directory to save results.}

\item{fname}{unique file name.}
}
\description{
Depending on the model and response type a \code{k_ontram}, \code{Polr} or \code{glm} model is fit.
The function handles ordinal and binary responses.
}
