% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ref_lossrps.R
\name{fit_ref_lossrps}
\alias{fit_ref_lossrps}
\title{Fit unconditional (SI) or linear shift model (SI-LSx) by minimizing the RPS}
\usage{
fit_ref_lossrps(
  mod = c("sils", "si"),
  fml,
  tab_dat,
  ridx,
  splits = 6,
  bs = ncol(tab_dat),
  lr = 0.005,
  decay = 0.001,
  optimizer = optimizer_adam(learning_rate = lr, decay = decay),
  epochs = 800,
  out_dir,
  fname
)
}
\arguments{
\item{fml}{model formula (e.g. \code{y ~ 1} for fitting the unconditional model).}

\item{tab_dat}{data frame of tabular data (including response variable). Response variable needs to be ordered
(also for binary responses).}

\item{ridx}{row indices used for each split (output from \code{get_ridx}).}

\item{out_dir}{directory to save results.}

\item{fname}{unique file name.}
}
\description{
Fit unconditional (SI) or linear shift model (SI-LSx) by minimizing the RPS
}
