% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generator}
\alias{generator}
\title{Data generator}
\usage{
generator(
  mod = c("silscs", "sics", "cils", "ci"),
  im,
  x = NULL,
  y,
  batch_size = 32,
  shuffle = TRUE
)
}
\description{
Data generator that feeds data real-time to model.
}
