% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_acc}
\alias{get_acc}
\title{Calculate accuracy}
\usage{
get_acc(cdf, y_true)
}
\description{
Calculate accuracy
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.6, 1,
                           0.5, 0.7, 1,
                           0.3, 0.4, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_acc(cdf = cdf, y_true = y_true)
}
