% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_auc}
\alias{get_auc}
\title{Calculate area under the ROC curve}
\usage{
get_auc(cdf, y_true, cutoff = 3)
}
\description{
Calculate area under the ROC curve
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.6, 1,
                           0.5, 0.7, 1,
                           0.3, 0.4, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_auc(cdf = cdf, y_true = y_true, cutoff = 2)
}
