% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_avg_cal}
\alias{get_avg_cal}
\title{Calculate mean calibration intercept and slope across classes and ensemble members}
\usage{
get_avg_cal(lys_cdf, y_true, weights = rep(1, length(lys_cdf)))
}
\arguments{
\item{lys_cdf}{list of CDFs (e.g. CDFs of ensemble members).}
}
\description{
Calculate mean calibration intercept and slope across classes and ensemble members
}
\examples{
cdf1 <- data.frame(matrix(c(0.2, 0.6, 1,
                            0.5, 0.7, 1,
                            0.3, 0.4, 1),
                          nrow = 3, byrow = T))
cdf2 <- data.frame(matrix(c(0.1, 0.8, 1,
                            0.3, 0.4, 1,
                            0.2, 0.6, 1),
                          nrow = 3, byrow = T))
lys_cdf <- list(cdf1, cdf2)
y_true <- data.frame(matrix(c(0, 1, 0,
                              0, 1, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_avg_cal(lys_cdf = lys_cdf, y_true = y_true)
}
