% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_binacc}
\alias{get_binacc}
\title{Calculate binary accuracy}
\usage{
get_binacc(cdf, y_true, cutoff = 3, contribution = FALSE)
}
\description{
Calculate binary accuracy
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.6, 1,
                           0.5, 0.7, 1,
                           0.1, 0.8, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_binacc(cdf = cdf, y_true = y_true, cutoff = 2)
}
