% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_bins}
\alias{get_bins}
\title{Get predicted and observed probabilities}
\usage{
get_bins(cdf, y_true, bins = 10)
}
\description{
Get predicted and observed probabilities
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.6, 1,
                           0.5, 0.7, 1,
                           0.1, 0.8, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_bins(cdf = cdf, y_true = y_true, bins = 10)
}
