% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_brier_perclass}
\alias{get_brier_perclass}
\title{Calculate brier score per class}
\usage{
get_brier_perclass(cdf, y_true)
}
\description{
Calculate brier score per class
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.6, 1,
                           0.5, 0.7, 1,
                           0.3, 0.4, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_brier_perclass(cdf = cdf, y_true = y_true)
}
