% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_bs}
\alias{get_bs}
\title{Applies a function to a bootstrap sample of all CDFs within a split B times
and returns all values.}
\usage{
get_bs(
  lys_cdf_all,
  y_true_all,
  fun,
  B = 1000,
  weights = rep(list(rep(1, length(lys_cdf_all[[1]]))), length(lys_cdf_all)),
  ...
)
}
\arguments{
\item{B}{number of bootstrap samples that should be drawn per split.}

\item{weights}{list of weights. Each element contains the weights per split as numeric vector.}

\item{...}{additional arguments to \code{fun}.}
}
\description{
Applies a function to a bootstrap sample of all CDFs within a split B times
and returns all values.
}
