% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_ensemble}
\alias{get_ensemble}
\title{Ensemble}
\usage{
get_ensemble(lys_cdf, type = c("linear", "log-linear", "trafo"))
}
\description{
Ensemble
}
\examples{
cdf1 <- data.frame(matrix(c(0.1, 0.2, 1,
                            0.3, 0.7, 1,
                            0.5, 0.7, 1),
                          nrow = 3, byrow = TRUE))
cdf2 <- data.frame(matrix(c(0.2, 0.3, 1,
                            0.5, 0.8, 1,
                            0.2, 0.9, 1),
                          nrow = 3, byrow = TRUE))
cdf3 <- data.frame(matrix(c(0.7, 0.8, 1,
                            0.6, 0.8, 1,
                            0.2, 0.8, 1),
                          nrow = 3, byrow = TRUE))
lys_cdf <- list(cdf1, cdf2, cdf3)

get_ensemble(lys_cdf = lys_cdf, type = "linear")
get_ensemble(lys_cdf = lys_cdf, type = "log-linear")
get_ensemble(lys_cdf = lys_cdf, type = "trafo")
}
