% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_indiv_metrics}
\alias{get_indiv_metrics}
\title{Evaluation of single ensemble members}
\usage{
get_indiv_metrics(
  lys_cdf,
  y_true,
  metrics = c("all", "acc", "binacc", "eacc", "ebinacc", "nll", "binnll", "rps",
    "cint", "cslope", "brier", "auc", "eauc", "qwk", "eqwk"),
  cutoff = 3,
  p = 2
)
}
\arguments{
\item{lys_cdf}{list of CDFs (e.g. CDFs of ensemble members).}

\item{y_true}{one-hot encoded true \code{y}.}

\item{metrics}{metrics to evaluate.}

\item{cutoff}{cutoff used to evaluate binary metrics.}

\item{p}{weighting scheme used for qwk.}
}
\description{
Evaluation of single ensemble members
}
