% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_metrics}
\alias{get_metrics}
\title{Evaluation of ensemble}
\usage{
get_metrics(
  lys_cdf,
  y_true,
  type = c("all", "linear", "log-linear", "trafo", "avg"),
  metrics = c("all", "acc", "binacc", "eacc", "ebinacc", "nll", "binnll", "rps",
    "cint", "cslope", "brier", "auc", "eauc", "qwk", "eqwk"),
  topk = TRUE,
  order_metric = c("nll", "rps"),
  lys_cdf_val = NULL,
  y_true_val = NULL,
  cutoff = 3,
  p = 2,
  weights = rep(1, length(lys_cdf))
)
}
\arguments{
\item{lys_cdf}{list of CDFs (e.g. CDFs of ensemble members).}

\item{y_true}{one-hot encoded true \code{y}.}

\item{type}{ensemble technique.}

\item{metrics}{metrics to evaluate.}

\item{topk}{logical. Whether to evaluate top-k ensembles from \code{1:K} based on validation loss.}

\item{order_metric}{metric used to evaluate and order validation CDFs.}

\item{lys_cdf_val}{list of CDFs.}

\item{y_true_val}{one-hot encoded true \code{y} of validation set.}

\item{cutoff}{cutoff used to evaluate binary metrics.}

\item{p}{weighting scheme used for qwk.}

\item{weights}{vector of weights for ensemble members.}
}
\description{
Evaluation of ensemble
}
