% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_metrics_allspl}
\alias{get_metrics_allspl}
\title{Evaluation of all splits}
\usage{
get_metrics_allspl(
  lys_cdf_all,
  y_true_all,
  type = c("all", "linear", "log-linear", "trafo", "avg"),
  metrics = c("all", "acc", "binacc", "eacc", "ebinacc", "nll", "binnll", "rps",
    "cint", "cslope", "brier", "auc", "eauc", "qwk", "eqwk"),
  topk = TRUE,
  order_metric = c("nll", "rps"),
  lys_cdf_val_all = NULL,
  y_true_val_all = NULL,
  cutoff = 3,
  p = 2,
  weights = matrix(1, nrow = length(lys_cdf_all), ncol = length(lys_cdf_all[[1]]))
)
}
\arguments{
\item{lys_cdf_all}{list of all CDFs (e.g. output of \code{\link{list_cdfs}})}

\item{y_true_all}{list of one-hot encoded true \code{y} for all splits.}

\item{type}{ensemble technique.}

\item{metrics}{metrics to evaluate.}

\item{topk}{logical. Whether to evaluate top-k ensembles from \code{1:K} based on validation loss.}

\item{order_metric}{metric used to evaluate and order validation CDFs.}

\item{lys_cdf_val_all}{list of all CDFs.}

\item{y_true_val_all}{list of one-hot encoded true \code{y} of validation set for all splits.}

\item{cutoff}{cutoff used to evaluate binary metrics.}

\item{p}{weighting scheme used for qwk.}

\item{weights}{matrix of weights (cols) per split (rows).}
}
\description{
Evaluation of all splits
}
