% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_order}
\alias{get_order}
\title{Ordered indices of CDFs based on NLL or RPS}
\usage{
get_order(lys_cdf_val, y_true_val, order_metric = c("nll", "rps"))
}
\description{
Ordered indices of CDFs based on NLL or RPS
}
\examples{
cdf1 <- data.frame(matrix(c(0.1, 0.2, 1,
                            0.3, 0.7, 1,
                            0.5, 0.7, 1),
                          nrow = 3, byrow = TRUE))
cdf2 <- data.frame(matrix(c(0.2, 0.3, 1,
                            0.5, 0.8, 1,
                            0.2, 0.9, 1),
                          nrow = 3, byrow = TRUE))
cdf3 <- data.frame(matrix(c(0.7, 0.8, 1,
                            0.6, 0.8, 1,
                            0.2, 0.8, 1),
                          nrow = 3, byrow = TRUE))
lys_cdf <- list(cdf1, cdf2, cdf3)
y_true <- data.frame(matrix(c(0, 0, 1,
                              1, 0, 0,
                              0, 1, 0),
                           nrow = 3, byrow = TRUE))

get_order(lys_cdf_val = lys_cdf, y_true = y_true, order_metric = "rps")
get_order(lys_cdf_val = lys_cdf, y_true = y_true, order_metric = "nll")
}
