% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_qwk}
\alias{get_qwk}
\title{Calculate quadratic weighted kappa}
\usage{
get_qwk(cdf, y_true, p = 2, contribution = FALSE)
}
\description{
Calculate quadratic weighted kappa
}
\examples{
cdf <- data.frame(matrix(c(0.2, 0.3, 1,
                           0.5, 0.7, 1,
                           0.3, 0.4, 1),
                         nrow = 3, byrow = T))
y_true <- data.frame(matrix(c(0, 1, 0,
                              1, 0, 0,
                              0, 0, 1),
                            nrow = 3, byrow = T))
get_qwk(cdf = cdf, y_true = y_true, p = 2)
}
