% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_single_bs}
\alias{get_single_bs}
\title{Applies a function to a bootstrap sample of all CDFs in a list and returns their mean}
\usage{
get_single_bs(lys_cdf, y_true, fun, weights = rep(1, length(lys_cdf)), ...)
}
\arguments{
\item{fun}{function that is applied to a bootstrap sample. Must take as
arguments \code{cdf} and \code{y_true}.}

\item{...}{additional arguments to \code{fun}.}
}
\description{
Applies a function to a bootstrap sample of all CDFs in a list and returns their mean
}
