% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_single_bs_across_spl}
\alias{get_single_bs_across_spl}
\title{Applies a function to a bootstrap sample of all CDFs within a split and
returns their mean across splits.}
\usage{
get_single_bs_across_spl(
  lys_cdf_all,
  y_true_all,
  fun,
  weights = rep(list(rep(1, length(lys_cdf_all[[1]]))), length(lys_cdf_all)),
  ...
)
}
\arguments{
\item{weights}{list of weights. Each element contains the weights per split as numeric vector.}

\item{...}{additional arguments to \code{fun}.}
}
\description{
Applies a function to a bootstrap sample of all CDFs within a split and
returns their mean across splits.
}
