% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_w}
\alias{get_w}
\title{Get weights per ensemble member}
\usage{
get_w(
  lys_cdf_val,
  y_true_val,
  type = c("linear", "log-linear", "trafo"),
  optim_metric = c("nll", "rps")
)
}
\description{
Get weights per ensemble member
}
\examples{
cdf1 <- data.frame(matrix(c(0.1, 0.2, 1,
                            0.3, 0.7, 1,
                            0.5, 0.7, 1),
                          nrow = 3, byrow = TRUE))
cdf2 <- data.frame(matrix(c(0.2, 0.3, 1,
                            0.5, 0.8, 1,
                            0.2, 0.9, 1),
                          nrow = 3, byrow = TRUE))
cdf3 <- data.frame(matrix(c(0.7, 0.8, 1,
                            0.6, 0.8, 1,
                            0.2, 0.8, 1),
                          nrow = 3, byrow = TRUE))
lys_cdf <- list(cdf1, cdf2, cdf3)
y_true <- data.frame(matrix(c(0, 0, 1,
                              1, 0, 0,
                              0, 1, 0),
                           nrow = 3, byrow = TRUE))

get_w(lys_cdf_val = lys_cdf, y_true_val = y_true, type = "trafo", optim_metric = "nll")
}
