% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-results.R
\name{list_cdfs}
\alias{list_cdfs}
\title{Load CDFs of all splits and ensemble members}
\usage{
list_cdfs(
  in_dir,
  fname,
  splits,
  ensembles = NULL,
  type = c("test", "val", "train")
)
}
\description{
Load CDFs of all splits and ensemble members
}
