% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-results.R
\name{list_lors}
\alias{list_lors}
\title{Load log-odds ratios of all splits and ensemble members}
\usage{
list_lors(in_dir, fname, splits, ensembles = NULL)
}
\description{
Load log-odds ratios of all splits and ensemble members
}
