% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-results.R
\name{list_terms}
\alias{list_terms}
\title{Load raw terms on the scale of \code{h} of all splits and ensemble members}
\usage{
list_terms(
  in_dir,
  fname,
  splits,
  ensembles = NULL,
  type = c("test", "val", "train")
)
}
\description{
Load raw terms on the scale of \code{h} of all splits and ensemble members
}
