% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-results.R
\name{list_trafos}
\alias{list_trafos}
\title{Load predictions on the scale of \code{h} of all splits and ensemble members}
\usage{
list_trafos(
  in_dir,
  fname,
  splits,
  ensembles = NULL,
  type = c("test", "val", "train")
)
}
\description{
Load predictions on the scale of \code{h} of all splits and ensemble members
}
