% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-results.R
\name{load_hists}
\alias{load_hists}
\title{Load histories of all splits and ensemble members}
\usage{
load_hists(in_dir, fname, splits, ensembles)
}
\description{
Load histories of all splits and ensemble members
}
