% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_hists}
\alias{plot_hists}
\title{Plot histories}
\usage{
plot_hists(
  hists,
  ylab = "",
  title = "",
  nrow = NULL,
  ncol = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{hists}{output of \code{\link{load_hists}}.}
}
\description{
Plot histories
}
