% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot evaluation metrics}
\usage{
plot_metrics(
  met,
  indiv_met = NULL,
  metrics = c("all", "acc", "eacc", "nll", "binnll", "rps", "cint", "cslope", "brier",
    "auc", "eauc", "qwk"),
  facet_labs = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{met}{output of \code{\link{get_metrics_allspl}}.}

\item{indiv_met}{output of \code{\link{get_indiv_metrics_allspl}}.
If not \code{NULL} rugs of performance of single ensemble members will be added to the plot.}

\item{metrics}{which metrics should be plotted.}

\item{facet_labs}{labels used for facets.}
}
\description{
Plot evaluation metrics
}
