% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-load-idx.R
\name{save_ridx}
\alias{save_ridx}
\title{Save row indices of random splits}
\usage{
save_ridx(nsplts, prptest, prpval, tab_dat, fml, out_dir, fname)
}
\arguments{
\item{prptest}{proportion of observations used for testing.}

\item{prpval}{proportion of observations used for validation.}

\item{tab_dat}{data frame containing at least the response variable.}

\item{fml}{model formula (only used to identify the response variable).}

\item{nsplits}{number of splits.}
}
\description{
Saves row indices used for training, validating and testing for each split as a data frame.
Ensures that all outcome classes are present in the training set.
}
