% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{warm_mod}
\alias{warm_mod}
\title{Initialize weights for simple intercept and linear shift term}
\usage{
warm_mod(m, mod = c("silscs", "sics", "cils"), x = NULL, y, binary = FALSE)
}
\description{
Initialize weights for simple intercept and linear shift term
}
\examples{
set.seed(2022)
nn <- function(output_shape = NULL, mbl = FALSE, input_shape = dim(im)[2:4], ...) {
  m <- keras_model_sequential() \%>\%
    layer_conv_2d(filters = 16,
                  input_shape = input_shape, kernel_size = c(3, 3), activation = "relu") \%>\%
    layer_max_pooling_2d(pool_size = c(2, 2)) \%>\%
    layer_conv_2d(filters = 32, kernel_size = c(3, 3), padding = "same", activation = "relu") \%>\%
    layer_max_pooling_2d(pool_size = c(2, 2)) \%>\%
    layer_flatten() \%>\%
    layer_dense(units = 50, activation = "relu") \%>\%
    layer_dense(output_shape, activation = "linear", use_bias = FALSE) \%>\%
    {if (mbl){
      ontram::layer_trafo_intercept()(.)
    } else .
    }
  return(m)
}

mnist <- dataset_mnist()
c(c(x_train, y_train), c(x_test, y_test)) \%<-\% mnist

m <- get_model(mod = "silscs", y_dim = 10L, x_dim = 2L, nn = nn, input_shape = c(28, 28, 1))

x <- model.frame(data.frame(x1 = rnorm(100), x2 = rnorm(100)))
y <- to_categorical(y_train[1:100])

get_weights(m$mod_baseline)
get_weights(m$list_of_shift_models[[0]])
m <- warm_mod(m, mod = "silscs", x = x, y = y, binary = FALSE)
get_weights(m$mod_baseline)
get_weights(m$list_of_shift_models[[0]])
}
